[refeactorable] Have a generated oject function that return the name of a field #181 - Generate file
