+++ go/models/model_file_gong.go
@@ -10,385 +10,6 @@ import (
 	"strings"
 )
 
-const ModelGongFileTemplate = `// generated by ModelGongFileTemplate
-package models
-
-import (
-	"fmt"
-	"log"
-	"os"
-	"path"
-	"regexp"
-	"sort"
-	"strings"
-)
-
-// swagger:ignore
-type __void any
-
-// needed for creating set of instances in the stage
-var __member __void
-
-// GongStructInterface is the interface met by GongStructs
-// It allows runtime reflexion of instances (without the hassle of the "reflect" package)
-type GongStructInterface interface {
-	GetName() (res string)
-	GetFields() (res []string)
-	GetFieldStringValue(fieldName string) (res string)
-}
-
-// StageStruct enables storage of staged instances
-// swagger:ignore
-type StageStruct struct { // insertion point for definition of arrays registering instances{{` + string(rune(ModelGongStructInsertionArrayDefintion)) + `}}
-	AllModelsStructCreateCallback AllModelsStructCreateInterface
-
-	AllModelsStructDeleteCallback AllModelsStructDeleteInterface
-
-	BackRepo BackRepoInterface
-
-	// if set will be called before each commit to the back repo
-	OnInitCommitCallback          OnInitCommitInterface
-	OnInitCommitFromFrontCallback OnInitCommitInterface
-	OnInitCommitFromBackCallback  OnInitCommitInterface
-
-	// store the number of instance per gongstruct
-	Map_GongStructName_InstancesNb map[string]int
-}
-
-type OnInitCommitInterface interface {
-	BeforeCommit(stage *StageStruct)
-}
-
-type BackRepoInterface interface {
-	Commit(stage *StageStruct)
-	Checkout(stage *StageStruct)
-	Backup(stage *StageStruct, dirPath string)
-	Restore(stage *StageStruct, dirPath string)
-	BackupXL(stage *StageStruct, dirPath string)
-	RestoreXL(stage *StageStruct, dirPath string)
-	// insertion point for Commit and Checkout signatures{{` + string(rune(ModelGongStructInsertionCommitCheckout)) + `}}
-	GetLastCommitFromBackNb() uint
-	GetLastPushFromFrontNb() uint
-}
-
-// swagger:ignore instructs the gong compiler (gongc) to avoid this particular struct
-var Stage StageStruct = StageStruct{ // insertion point for array initiatialisation{{` + string(rune(ModelGongStructInsertionArrayInitialisation)) + `}}
-	// end of insertion point
-	Map_GongStructName_InstancesNb: make(map[string]int),
-}
-
-func (stage *StageStruct) Commit() {
-	if stage.BackRepo != nil {
-		stage.BackRepo.Commit(stage)
-	}
-
-	// insertion point for computing the map of number of instances per gongstruct{{` + string(rune(ModelGongStructInsertionComputeNbInstances)) + `}}
-
-}
-
-func (stage *StageStruct) Checkout() {
-	if stage.BackRepo != nil {
-		stage.BackRepo.Checkout(stage)
-	}
-
-	// insertion point for computing the map of number of instances per gongstruct{{` + string(rune(ModelGongStructInsertionComputeNbInstances)) + `}}
-
-}
-
-// backup generates backup files in the dirPath
-func (stage *StageStruct) Backup(dirPath string) {
-	if stage.BackRepo != nil {
-		stage.BackRepo.Backup(stage, dirPath)
-	}
-}
-
-// Restore resets Stage & BackRepo and restores their content from the restore files in dirPath
-func (stage *StageStruct) Restore(dirPath string) {
-	if stage.BackRepo != nil {
-		stage.BackRepo.Restore(stage, dirPath)
-	}
-}
-
-// backup generates backup files in the dirPath
-func (stage *StageStruct) BackupXL(dirPath string) {
-	if stage.BackRepo != nil {
-		stage.BackRepo.BackupXL(stage, dirPath)
-	}
-}
-
-// Restore resets Stage & BackRepo and restores their content from the restore files in dirPath
-func (stage *StageStruct) RestoreXL(dirPath string) {
-	if stage.BackRepo != nil {
-		stage.BackRepo.RestoreXL(stage, dirPath)
-	}
-}
-
-// insertion point for cumulative sub template with model space calls{{` + string(rune(ModelGongStructInsertionStageFunctions)) + `}}
-// swagger:ignore
-type AllModelsStructCreateInterface interface { // insertion point for Callbacks on creation{{` + string(rune(ModelGongStructInsertionCreateCallback)) + `}}
-}
-
-type AllModelsStructDeleteInterface interface { // insertion point for Callbacks on deletion{{` + string(rune(ModelGongStructInsertionDeleteCallback)) + `}}
-}
-
-func (stage *StageStruct) Reset() { // insertion point for array reset{{` + string(rune(ModelGongStructInsertionArrayReset)) + `}}
-}
-
-func (stage *StageStruct) Nil() { // insertion point for array nil{{` + string(rune(ModelGongStructInsertionArrayNil)) + `}}
-}
-
-const marshallRes = ` + "`" + `package {{PackageName}}
-
-import (
-	"time"
-
-	"{{ModelsPackageName}}"
-)
-
-func init() {
-	var __Dummy_time_variable time.Time
-	_ = __Dummy_time_variable
-	InjectionGateway["{{databaseName}}"] = {{databaseName}}Injection
-}
-
-// {{databaseName}}Injection will stage objects of database "{{databaseName}}"
-func {{databaseName}}Injection() {
-
-	// Declaration of instances to stage{{Identifiers}}
-
-	// Setup of values{{ValueInitializers}}
-
-	// Setup of pointers{{PointersInitializers}}
-}
-
-` + "`" + `
-
-const IdentifiersDecls = ` + "`" + `
-	{{Identifier}} := (&models.{{GeneratedStructName}}{Name: "{{GeneratedFieldNameValue}}"}).Stage()` + "`" + `
-
-const StringInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}} = ` +
-	"`" + " + \"" + "`" + `"` + ` + ` + "`" +
-	`{{GeneratedFieldNameValue}}` + "`" + ` + "` + "`" + `"` + `
-
-const StringEnumInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}} = {{GeneratedFieldNameValue}}` + "`" + `
-
-const NumberInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}} = {{GeneratedFieldNameValue}}` + "`" + `
-
-const PointerFieldInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}} = {{GeneratedFieldNameValue}}` + "`" + `
-
-const SliceOfPointersFieldInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}} = append({{Identifier}}.{{GeneratedFieldName}}, {{GeneratedFieldNameValue}})` + "`" + `
-
-const TimeInitStatement = ` + "`" + `
-	{{Identifier}}.{{GeneratedFieldName}}, _ = time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", "{{GeneratedFieldNameValue}}")` + "`" + `
-
-// Marshall marshall the stage content into the file as an instanciation into a stage
-func (stage *StageStruct) Marshall(file *os.File, modelsPackageName, packageName string) {
-
-	name := file.Name()
-
-	if !strings.HasSuffix(name, ".go") {
-		log.Fatalln(name + " is not a go filename")
-	}
-
-	log.Println("filename of marshall output  is " + name)
-
-	res := marshallRes
-	res = strings.ReplaceAll(res, "{{databaseName}}", strings.ReplaceAll(path.Base(name), ".go", ""))
-	res = strings.ReplaceAll(res, "{{PackageName}}", packageName)
-	res = strings.ReplaceAll(res, "{{ModelsPackageName}}", modelsPackageName)
-
-	// map of identifiers
-	// var StageMapDstructIds map[*Dstruct]string
-	identifiersDecl := ""
-	initializerStatements := ""
-	pointersInitializesStatements := ""
-
-	id := ""
-	decl := ""
-	setValueField := ""
-
-	// insertion initialization of objects to stage{{` + string(rune(ModelGongStructInsertionUnmarshallDeclarations)) + `}}
-	// insertion initialization of objects to stage{{` + string(rune(ModelGongStructInsertionUnmarshallPointersInitializations)) + `}}
-	res = strings.ReplaceAll(res, "{{Identifiers}}", identifiersDecl)
-	res = strings.ReplaceAll(res, "{{ValueInitializers}}", initializerStatements)
-	res = strings.ReplaceAll(res, "{{PointersInitializers}}", pointersInitializesStatements)
-
-	fmt.Fprintln(file, res)
-}
-
-// unique identifier per struct
-func generatesIdentifier(gongStructName string, idx int, instanceName string) (identifier string) {
-
-	identifier = instanceName
-	// Make a Regex to say we only want letters and numbers
-	reg, err := regexp.Compile("[^a-zA-Z0-9]+")
-	if err != nil {
-		log.Fatal(err)
-	}
-	processedString := reg.ReplaceAllString(instanceName, "_")
-
-	identifier = fmt.Sprintf("__%s__%06d_%s", gongStructName, idx, processedString)
-
-	return
-}
-
-// insertion point of functions that provide maps for reverse associations{{` + string(rune(ModelGongStructInsertionReverseAssociationsMaps)) + `}}
-
-// Gongstruct is the type paramter for generated generic function that allows
-// - access to staged instances
-// - navigation between staged instances by going backward association links between gongstruct
-// - full refactoring of Gongstruct identifiers / fields
-type Gongstruct interface {
-	// insertion point for generic types
-	{{` + string(rune(ModelGongStructInsertionGenericGongstructTypes)) + `}}
-}
-
-type GongstructSet interface {
-	map[any]any |
-		// insertion point for generic types{{` + string(rune(ModelGongStructInsertionGenericGongSetTypes)) + `}}
-		map[*any]any // because go does not support an extra "|" at the end of type specifications
-}
-
-type GongstructMapString interface {
-	map[any]any |
-		// insertion point for generic types{{` + string(rune(ModelGongStructInsertionGenericGongMapTypes)) + `}}
-		map[*any]any // because go does not support an extra "|" at the end of type specifications
-}
-
-// GongGetSet returns the set staged GongstructType instances
-// it is usefull because it allows refactoring of gong struct identifier
-func GongGetSet[Type GongstructSet]() *Type {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get functions{{` + string(rune(ModelGongStructInsertionGenericGetSetFunctions)) + `}}
-	default:
-		return nil
-	}
-}
-
-// GongGetMap returns the map of staged GongstructType instances
-// it is usefull because it allows refactoring of gong struct identifier
-func GongGetMap[Type GongstructMapString]() *Type {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get functions{{` + string(rune(ModelGongStructInsertionGenericGetMapFunctions)) + `}}
-	default:
-		return nil
-	}
-}
-
-// GetGongstructInstancesSet returns the set staged GongstructType instances
-// it is usefull because it allows refactoring of gongstruct identifier
-func GetGongstructInstancesSet[Type Gongstruct]() *map[*Type]any {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get functions{{` + string(rune(ModelGongStructInsertionGenericInstancesSetFunctions)) + `}}
-	default:
-		return nil
-	}
-}
-
-// GetGongstructInstancesMap returns the map of staged GongstructType instances
-// it is usefull because it allows refactoring of gong struct identifier
-func GetGongstructInstancesMap[Type Gongstruct]() *map[string]*Type {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get functions{{` + string(rune(ModelGongStructInsertionGenericInstancesMapFunctions)) + `}}
-	default:
-		return nil
-	}
-}
-
-// GetAssociationName is a generic function that returns an instance of Type
-// where each association is filled with an instance whose name is the name of the association
-//
-// This function can be handy for generating navigation function that are refactorable
-func GetAssociationName[Type Gongstruct]() *Type {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for instance with special fields{{` + string(rune(ModelGongStructInsertionGenericGetAssociationNameFunctions)) + `}}
-	default:
-		return nil
-	}
-}
-
-// GetPointerReverseMap allows backtrack navigation of any Start.Fieldname
-// associations (0..1) that is a pointer from one staged Gongstruct (type Start)
-// instances to another (type End)
-//
-// The function provides a map with keys as instances of End and values to arrays of *Start
-// the map is construed by iterating over all Start instances and populationg keys with End instances
-// and values with slice of Start instances
-func GetPointerReverseMap[Start, End Gongstruct](fieldname string) map[*End][]*Start {
-	var ret Start
-
-	switch any(ret).(type) {
-	// insertion point of functions that provide maps for reverse associations{{` + string(rune(ModelGongStructInsertionGenericReversePointerAssociationsMaps)) + `}}
-	}
-	return nil
-}
-
-// GetSliceOfPointersReverseMap allows backtrack navigation of any Start.Fieldname
-// associations (0..N) between one staged Gongstruct instances and many others
-//
-// The function provides a map with keys as instances of End and values to *Start instances
-// the map is construed by iterating over all Start instances and populating keys with End instances
-// and values with the Start instances
-func GetSliceOfPointersReverseMap[Start, End Gongstruct](fieldname string) map[*End]*Start {
-	var ret Start
-
-	switch any(ret).(type) {
-	// insertion point of functions that provide maps for reverse associations{{` +
-	string(rune(ModelGongStructInsertionGenericReverseSliceOfPointersAssociationsMaps)) + `}}
-	}
-	return nil
-}
-
-// GetGongstructName returns the name of the Gongstruct
-// this can be usefull if one want program robust to refactoring
-func GetGongstructName[Type Gongstruct]() (res string) {
-
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get gongstruct name{{` + string(rune(ModelGongStructInsertionGenericGongstructName)) + `}}
-	}
-	return res
-}
-
-// GetFields return the array of the fields
-func GetFields[Type Gongstruct]() (res []string) {
-
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get gongstruct name{{` + string(rune(ModelGongStructInsertionGenericGetFields)) + `}}
-	}
-	return
-}
-
-func GetFieldStringValue[Type Gongstruct](instance Type, fieldName string) (res string) {
-	var ret Type
-
-	switch any(ret).(type) {
-	// insertion point for generic get gongstruct field value{{` + string(rune(ModelGongStructInsertionGenericGetFieldValues)) + `}}
-	}
-	return
-}
-
-// insertion point of enum utility functions{{` + string(rune(ModelGongEnumUtilityFunctions)) + `}}
-`
-
 //
 // insertion points are places where the code is
 // generated per gong struct
@@ -1255,7 +876,7 @@ func CodeGeneratorModelGong(
 	}
 
 	// substitutes {{<<insertionPerEnumId points>>}} stuff with generated code
-	for insertionPerEnumId := ModelGongEnumInsertionId(0); insertionPerEnumId < ModelGongEnumInsertionsNb; insertionPerEnumId++ {
+	for insertionPerEnumId := ModelGongEnumUtilityFunctions; insertionPerEnumId < ModelGongEnumInsertionsNb; insertionPerEnumId++ {
 		toReplace := "{{" + string(rune(insertionPerEnumId)) + "}}"
 		codeGO = strings.ReplaceAll(codeGO, toReplace, subEnumCodes[insertionPerEnumId])
 	}
