+++ test/go/tests/simple_test.go
@@ -190,18 +190,18 @@ func TestStageCallBack(t *testing.T) {
 		t.Errorf("got = %d; want %d", got, want)
 	}
 
-	fieldCoder := models.GongfieldCoder[models.Astruct]()
-	log.Println("Field namer code : ", fieldCoder.Name)
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Name))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Booleanfield))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherbooleanfield))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Intfield))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Floatfield))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.CFloatfield))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Associationtob))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherassociationtob_2))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anarrayofb))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherarrayofb))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Aenum))
-	log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.CEnum))
+	// fieldCoder := models.GongfieldCoder[models.Astruct]()
+	// log.Println("Field namer code : ", fieldCoder.Name)
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Name))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Booleanfield))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherbooleanfield))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Intfield))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Floatfield))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.CFloatfield))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Associationtob))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherassociationtob_2))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anarrayofb))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Anotherarrayofb))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.Aenum))
+	// log.Println("Field namer name : ", models.GongfieldName[models.Astruct](fieldCoder.CEnum))
 }
